generate random tweets script updated, changes to elasticsearch connection

+++ scripts/generate_random_tweets.py
@@ -8,6 +8,7 @@ import random
 import time
 from datetime import datetime
 import string
+from faker import Faker
 
 
 def get_random_lon_lat():
@@ -24,11 +25,13 @@ def generate_fake_tweet():
     twitter_time_format = '%a %b %d %H:%M:%S +0000 %Y'
     time_now = datetime.utcnow().strftime(twitter_time_format)
     timestamp = time.time()
+    f = Faker()
+    profile = f.profile()
     t = { 
             'id': int(timestamp*100000000),
             'created_at': time_now,
             'project': INDEX,
-            'text': generate_random_text(words=10),
+            'text': f.text(),
             'lang': 'en',
             'coordinates': {
                 'coordinates': get_random_lon_lat(),
@@ -36,22 +39,22 @@ def generate_fake_tweet():
                 },
             'timestamp_ms': str(int(timestamp*1000)),
             'user': {
-                'description': generate_random_text(words=4),
-                'screen_name': generate_random_text(words=1),
+                'description': f.text(),
+                'screen_name': profile['username'],
                 'id_str': str(int(timestamp*100000)),
                 'lang': 'en',
-                'name': generate_random_text(words=2),
-                'location': generate_random_text(words=1),
-                'time_zone': 'Eastern Time (US & Canada)',
+                'name': profile['name'],
+                'location': f.country(),
+                'time_zone': f.timezone(),
                 'geo_enabled': True,
                 },
             "entities": {
-                "hashtags": [{ "indices": [71, 75], "text": "random" }]
+                "hashtags": [{ "indices": [71, 75], "text": f.word() }]
                 },
             "place": {
-                "country_code": "US",
-                "country": "United States",
-                "full_name": generate_random_text(words=2),
+                "country_code": f.country_code(),
+                "country": f.country(),
+                "full_name": f.city() + ' ' + f.country(),
                 "average_location": get_random_lon_lat(),
                 "place_type": "city",
                 "id": generate_random_text(words=1),
@@ -61,8 +64,9 @@ def generate_fake_tweet():
 
     if INDEX == 'project_vaccine_sentiment':
         predictions = {0: 'pro-vaccine', 1: 'anti-vaccine', 2:'neutral'}
-        sentiment = {'sentiment': {str(MODEL.split('.')[0]): {'label': predictions[random.randint(0,2)], 'probability': random.random()}}}
-        t['meta'] = sentiment
+        sentiment = {str(MODEL.split('.')[0]): {'label': predictions[random.randint(0,2)], 'probability': random.random(), 'val': random.uniform(-1,1)}}
+        t['meta'] = {'sentiment': sentiment}
+        t['sentiment'] = sentiment
     return t
 
 
@@ -86,10 +90,14 @@ if __name__ == "__main__":
     INDEX = 'project_vaccine_sentiment'
     DOC_TYPE='tweet'
     MODEL='fasttext_v1'
-    NUM_TWEETS=20                             # num tweets to generate
+    NUM_TWEETS=200
 
     logger.info('Generating random tweets and index them to index {}'.format(INDEX))
 
-    es_client.delete_index(INDEX)
-    es_client.create_index(INDEX)
+    if INDEX in es_client.list_templates():
+        yes_no = input('Index {} already exists for. Delete an recreate? (y/n)'.format(INDEX))
+        if yes_no == 'y':
+            es_client.delete_template('project')
+            es_client.delete_index(INDEX)
+            es_client.create_index(INDEX)
     main()
